refactor: address service dependency injection
