+++ social-network/src/main/java/com/levi9/socialnetwork/Service/AddressService.java
@@ -1,63 +1,23 @@
 package com.levi9.socialnetwork.Service;
 
-import java.util.HashMap;
 import java.util.Map;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import com.levi9.socialnetwork.Exception.ResourceNotFoundException;
 import com.levi9.socialnetwork.Model.Address;
-import com.levi9.socialnetwork.Repository.AddressRepository;
-import com.levi9.socialnetwork.Repository.UserRepository;
 
-@Service
-public class AddressService {
-	
-	
-	@Autowired
-	private AddressRepository addressRepository;
-	
-	public java.util.List<Address> getAllAddresses(){
-		return this.addressRepository.findAll();
-	}
-		
-	public ResponseEntity<Address> getAddressById(Long addressId)
-		throws ResourceNotFoundException{
-		Address address = addressRepository.findById(addressId).orElseThrow(() -> new ResourceNotFoundException("Address not found for this id :: " + addressId));
-	 return ResponseEntity.ok().body(address);
-	}
-	
+public interface AddressService {
 
-	public Address createAddress(Address address) {
-		return addressRepository.save(address);
-	}
+	public java.util.List<Address> getAllAddresses();
 	
-
-	public ResponseEntity<Address> updateAddress(Long addressId,
-			 @RequestBody Address addressDetails) throws ResourceNotFoundException {
-		Address address = addressRepository.findById(addressId)
-				.orElseThrow(() -> new ResourceNotFoundException("Address not found for this id :: " + addressId));
-		
-		address.setCountry(addressDetails.getCountry());
-		address.setCity(addressDetails.getCity());
-		address.setStreet(addressDetails.getStreet());
-		address.setNumber(addressDetails.getNumber());
+	public ResponseEntity<Address> getAddressById(Long addressId) throws ResourceNotFoundException;
 	
-		final Address updatedAddress = addressRepository.save(address);
-		return ResponseEntity.ok(updatedAddress);
-	}
+	public Address createAddress(Address address);
 	
-	public Map<String, Boolean> deleteAddress(Long addressId)
-			throws ResourceNotFoundException {
-		Address address = addressRepository.findById(addressId)
-				.orElseThrow(() -> new ResourceNotFoundException("Address not found for this id :: " + addressId));
-
-		addressRepository.delete(address);
-		Map<String, Boolean> response = new HashMap<>();
-		response.put("deleted", Boolean.TRUE);
-		return response;
-	}
+	public ResponseEntity<Address> updateAddress(Long addressId,@RequestBody Address addressDetails) throws ResourceNotFoundException;
+	
+	public Map<String, Boolean> deleteAddress(Long addressId) throws ResourceNotFoundException;
+	
 }
